fix(police): close upgrade-window race and state gaps
