+++ packages/surface-daemon-supervisor/README.md
@@ -6,8 +6,9 @@ to spawn, watch, and recycle a surface daemon it does *not* run in — the mirro
 daemon, so it is never a staleKey root. Beside the endpoint state machine it
 carries the **convergence kit** — the policy-driven answer to "the running daemon
 is not the one I shipped: detect it, decide, converge it." Depends only on
-`@kolu/surface-daemon` (the daemon-half twin) and `@kolu/surface` for the frozen
-control-core transport; no app package.
+`@kolu/surface-daemon` (the daemon-half twin), `@kolu/surface` for the frozen
+control-core transport, and `ts-pattern` for exhaustive policy dispatch; no app
+package.
 
 ```ts
 import {
