+++ packages/surface-daemon-supervisor/package.json
@@ -17,7 +17,8 @@
   },
   "dependencies": {
     "@kolu/surface": "workspace:*",
-    "@kolu/surface-daemon": "workspace:*"
+    "@kolu/surface-daemon": "workspace:*",
+    "ts-pattern": "^5.9.0"
   },
   "devDependencies": {
     "@babel/parser": "^7.29.2",
